The settings library can be symlinked to the common libraries dir

+++ create-library-symlinks.sh
@@ -18,8 +18,10 @@ function symlink-library {
   echo "- - -"
 
   for entry in $LIBRARIES_DIR/$name*; do
-    echo "- removing symlink: $entry"
-    rm $entry
+    if [ -e "$entry" ]; then
+      echo "- removing symlink: $entry"
+      rm $entry
+    fi
   done
 
   echo "- creating symlinks"
